fix: rework the whole architecture to not use a 'plugin' system as it does not really work with Akairo
- added info command
- added help command
- rolled back got because of promise hanging
