+++ .gitignore
@@ -2,7 +2,6 @@
 *
 
 # ADD CODE
-!index.js
 !src/
 !src/**
 !test
