+++ src/core/constants.js
@@ -3,12 +3,6 @@
 exports.CoreEvents = {
     SENTRY_INITIALIZED : 'sentryInitialized',
 
-    PLUGIN_REGISTERED : 'pluginRegistered',
-    PLUGIN_LOADED     : 'pluginLoaded',
-
-    PLUGIN_BEFORE_LOAD : 'pluginBeforeLoad',
-    PLUGIN_AFTER_LOAD  : 'pluginAfterLoad',
-
     PROVIDER_REGISTERED  : 'providerRegistered',
     PROVIDER_INITIALIZED : 'providerInitialized',
