+++ src/utils/capitalize.js
@@ -1,0 +1,6 @@
+'use strict';
+
+module.exports = (string) => {
+
+    return string[0].toUpperCase() + string.slice(1);
+};
